user-auth-api/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
bcrypt==3.2.0
PyJWT==2.0.0
six==1.15.0